chore: adjust map legends for mobile
