+++ i18n/messages/en/common.json
@@ -165,6 +165,7 @@
 		"switch-locale": "Switch to {locale}"
 	},
 	"MapPage": {
+		"mapLegend": "Map-Legend",
 		"meta": {
 			"title": "Map"
 		},
